bawt/requirements.txt
discord.py
flask
aiosqlite
requests
